+++ mongod/structs.go
@@ -31,14 +31,19 @@ type AppRepository interface {
 	CreateArch(archName string, owner string, ctx context.Context) (interface{}, error)
 	ListArchs(ctx context.Context, owner string) ([]*model.Arch, error)
 	DeleteArch(id primitive.ObjectID, owner string, ctx context.Context) (int64, error)
-	CreateApp(appName string, logo string, description string, private bool, tuf bool, owner string, ctx context.Context) (interface{}, error)
+	CreateApp(appName string, logo string, description string, private bool, tuf bool, reports bool, owner string, ctx context.Context) (interface{}, error)
 	ListApps(ctx context.Context, owner string) ([]*model.App, error)
 	DeleteApp(id primitive.ObjectID, owner string, ctx context.Context) (int64, error)
-	UpdateApp(id primitive.ObjectID, appName string, logo string, tuf bool, description string, owner string, ctx context.Context) (interface{}, error)
+	UpdateApp(id primitive.ObjectID, appName string, logo string, tuf bool, description string, reports bool, owner string, ctx context.Context) (interface{}, error)
 	UpdateChannel(id primitive.ObjectID, paramValue string, owner string, ctx context.Context) (interface{}, error)
 	UpdatePlatform(id primitive.ObjectID, platformName string, updaters []model.Updater, owner string, ctx context.Context) (interface{}, error)
 	UpdateArch(id primitive.ObjectID, paramValue string, owner string, ctx context.Context) (interface{}, error)
 	DeleteSpecificArtifactOfApp(id primitive.ObjectID, ctxQuery map[string]interface{}, ctx context.Context, owner string) ([]string, bool, error)
+	GetAppByID(id primitive.ObjectID, requester string, ctx context.Context) (*model.App, error)
+	CreateReportKey(appID primitive.ObjectID, requester string, ctx context.Context) (string, error)
+	DeleteReportKey(appID primitive.ObjectID, requester string, ctx context.Context) (bool, error)
+	ListReportKeys(requester string, ctx context.Context) ([]*model.ReportKeyListItem, error)
+	RegenerateReportKey(appID primitive.ObjectID, requester string, ctx context.Context) (string, error)
 }
 
 type appRepository struct {